update EBLT particles
